draft changes for threads

+++ tests/test_wkmeans_bounds.cpp
@@ -12,15 +12,15 @@ using namespace kmeans;
 TEST_CASE("test_update_bounds", "[bounds]"){
     PPF_SIZE = 20;
     build_data();
-    WKmeans<float> test_wkmeans(pdfs.size(), N_CLUSTERS, EPSILON, pdfs, cdfs, ppfs, 10, 42);
+    WKmeans<double> test_wkmeans(pdfs.size(), N_CLUSTERS, EPSILON, pdfs, cdfs, ppfs, 10, 42);
     test_wkmeans.init_clusters();
     test_wkmeans.init_bounds();
     test_wkmeans.update_clusters();
-    std::vector<float> upper_bounds = test_wkmeans.get_upper_bounds();
-    std::vector<float> lower_bounds = test_wkmeans.get_lower_bounds();
+    std::vector<double> upper_bounds = test_wkmeans.get_upper_bounds();
+    std::vector<double> lower_bounds = test_wkmeans.get_lower_bounds();
     std::vector<int> cluster_assignments = test_wkmeans.get_cluster_assignments();
-    std::vector<PPF<float>> clusters = test_wkmeans.clusters;
-    std::vector<PPF<float>> new_clusters = test_wkmeans.get_new_clusters();
+    std::vector<PPF<double>> clusters = test_wkmeans.clusters;
+    std::vector<PPF<double>> new_clusters = test_wkmeans.get_new_clusters();
     test_wkmeans.update_bounds();
     for(int i = 0; i < pdfs.size(); i ++){
         upper_bounds[i] += wasserstein_2(clusters[cluster_assignments[i]], new_clusters[cluster_assignments[i]]);
@@ -29,8 +29,8 @@ TEST_CASE("test_update_bounds", "[bounds]"){
 
     for(int i = 0; i < pdfs.size(); i++){
         for(int k = 0; k < N_CLUSTERS; k++){
-            float distance = wasserstein_2(clusters[k], new_clusters[k]);
-            lower_bounds[i*N_CLUSTERS + k] = std::max(lower_bounds[i*N_CLUSTERS + k] - distance, float{});
+            double distance = wasserstein_2(clusters[k], new_clusters[k]);
+            lower_bounds[i*N_CLUSTERS + k] = std::max(lower_bounds[i*N_CLUSTERS + k] - distance, double{});
         }
     }
     REQUIRE(lower_bounds == test_wkmeans.get_lower_bounds());
